Refactor BrainMamba: Fix bugs, add types, tests, and optimizations
- Refactored `SelectiveSSM` to correct dimension logic and use a reliable sequential scan.
- Replaced `.view()` with `.reshape()` in `BTMamba` and `BNMamba` to handle non-contiguous tensors.
- Added optional `input_dim` argument to `BTMamba` and `VariateEncoder` to properly initialize projection layers, preventing "untrained weights" issues with dynamic layers.
- Vectorized `construct_functional_connectivity` for significant performance improvement.
- Added comprehensive type hints across core modules.
- Added unit tests in `tests/` covering all modules.
- Updated `examples/simple_example.py` to handle BFloat16 types correctly.
- Cleaned up unused imports, variables, and comments.
- Removed broken parallel scan placeholder logic.

+++ examples/simple_example.py
@@ -168,13 +168,13 @@ def main():
         _, test_predicted = torch.max(test_logits, 1)
         
         # Compute accuracy
-        test_acc = accuracy_score(test_labels.cpu().float().numpy(), test_predicted.cpu().float().numpy())
+        test_acc = accuracy_score(test_labels.cpu().numpy(), test_predicted.cpu().numpy())
         print(f"Test Accuracy: {test_acc:.4f}")
         
         # Compute PR-AUC for binary classification
         if num_classes == 2:
             precision, recall, _ = precision_recall_curve(
-                test_labels.cpu().float().numpy(), test_probs[:, 1].cpu().float().numpy()
+                test_labels.cpu().numpy(), test_probs[:, 1].cpu().float().numpy()
             )
             pr_auc = auc(recall, precision)
             print(f"PR-AUC: {pr_auc:.4f}")
